задание 5(4).модуляция
